Refactor out helper for rendering forms
